Move ASTNode into separate rust file
