+++ src/context/ast_node.rs
@@ -1,0 +1,114 @@
+use crate::ast::*;
+
+#[derive(Debug, Clone, PartialEq)]
+pub enum ASTNode {
+    ArrayTypeName(ArrayTypeName),
+    Assignment(Assignment),
+    BinaryOperation(BinaryOperation),
+    Block(Block),
+    Conditional(Conditional),
+    ContractDefinition(ContractDefinition),
+    ElementaryTypeName(ElementaryTypeName),
+    ElementaryTypeNameExpression(ElementaryTypeNameExpression),
+    EmitStatement(EmitStatement),
+    EnumDefinition(EnumDefinition),
+    EnumValue(EnumValue),
+    EventDefinition(EventDefinition),
+    ErrorDefinition(ErrorDefinition),
+    ExpressionStatement(ExpressionStatement),
+    FunctionCall(FunctionCall),
+    FunctionCallOptions(FunctionCallOptions),
+    FunctionDefinition(FunctionDefinition),
+    FunctionTypeName(FunctionTypeName),
+    ForStatement(ForStatement),
+    Identifier(Identifier),
+    IdentifierPath(IdentifierPath),
+    IfStatement(IfStatement),
+    ImportDirective(ImportDirective),
+    IndexAccess(IndexAccess),
+    IndexRangeAccess(IndexRangeAccess),
+    InheritanceSpecifier(InheritanceSpecifier),
+    InlineAssembly(InlineAssembly),
+    Literal(Literal),
+    MemberAccess(MemberAccess),
+    NewExpression(NewExpression),
+    Mapping(Mapping),
+    ModifierDefinition(ModifierDefinition),
+    ModifierInvocation(ModifierInvocation),
+    OverrideSpecifier(OverrideSpecifier),
+    ParameterList(ParameterList),
+    PragmaDirective(PragmaDirective),
+    Return(Return),
+    RevertStatement(RevertStatement),
+    SourceUnit(SourceUnit),
+    StructDefinition(StructDefinition),
+    StructuredDocumentation(StructuredDocumentation),
+    TryStatement(TryStatement),
+    TryCatchClause(TryCatchClause),
+    TupleExpression(TupleExpression),
+    UnaryOperation(UnaryOperation),
+    UserDefinedTypeName(UserDefinedTypeName),
+    UserDefinedValueTypeDefinition(UserDefinedValueTypeDefinition),
+    UsingForDirective(UsingForDirective),
+    VariableDeclaration(VariableDeclaration),
+    VariableDeclarationStatement(VariableDeclarationStatement),
+    WhileStatement(WhileStatement),
+}
+
+impl ASTNode {
+    pub fn src(&self) -> Option<&str> {
+        match self {
+            ASTNode::ArrayTypeName(_) => None,
+            ASTNode::Assignment(node) => Some(&node.src),
+            ASTNode::BinaryOperation(node) => Some(&node.src),
+            ASTNode::Block(node) => Some(&node.src),
+            ASTNode::Conditional(node) => Some(&node.src),
+            ASTNode::ContractDefinition(node) => Some(&node.src),
+            ASTNode::ElementaryTypeName(_) => None,
+            ASTNode::ElementaryTypeNameExpression(node) => Some(&node.src),
+            ASTNode::EmitStatement(_) => None,
+            ASTNode::EnumDefinition(node) => Some(&node.src),
+            ASTNode::EnumValue(node) => Some(&node.src),
+            ASTNode::EventDefinition(node) => Some(&node.src),
+            ASTNode::ErrorDefinition(node) => Some(&node.src),
+            ASTNode::ExpressionStatement(_) => None,
+            ASTNode::FunctionCall(node) => Some(&node.src),
+            ASTNode::FunctionCallOptions(node) => Some(&node.src),
+            ASTNode::FunctionDefinition(node) => Some(&node.src),
+            ASTNode::FunctionTypeName(_) => None,
+            ASTNode::ForStatement(node) => Some(&node.src),
+            ASTNode::Identifier(node) => Some(&node.src),
+            ASTNode::IdentifierPath(node) => Some(&node.src),
+            ASTNode::IfStatement(node) => Some(&node.src),
+            ASTNode::ImportDirective(node) => Some(&node.src),
+            ASTNode::IndexAccess(node) => Some(&node.src),
+            ASTNode::IndexRangeAccess(node) => Some(&node.src),
+            ASTNode::InheritanceSpecifier(node) => Some(&node.src),
+            ASTNode::InlineAssembly(node) => Some(&node.src),
+            ASTNode::Literal(node) => Some(&node.src),
+            ASTNode::MemberAccess(node) => Some(&node.src),
+            ASTNode::NewExpression(node) => Some(&node.src),
+            ASTNode::Mapping(_) => None,
+            ASTNode::ModifierDefinition(node) => Some(&node.src),
+            ASTNode::ModifierInvocation(node) => Some(&node.src),
+            ASTNode::OverrideSpecifier(node) => Some(&node.src),
+            ASTNode::ParameterList(node) => Some(&node.src),
+            ASTNode::PragmaDirective(node) => Some(&node.src),
+            ASTNode::Return(node) => Some(&node.src),
+            ASTNode::RevertStatement(_) => None,
+            ASTNode::SourceUnit(_) => None,
+            ASTNode::StructDefinition(node) => Some(&node.src),
+            ASTNode::StructuredDocumentation(node) => Some(&node.src),
+            ASTNode::TryStatement(_) => None,
+            ASTNode::TryCatchClause(_) => None,
+            ASTNode::TupleExpression(node) => Some(&node.src),
+            ASTNode::UnaryOperation(node) => Some(&node.src),
+            ASTNode::UserDefinedTypeName(_) => None,
+            ASTNode::UserDefinedValueTypeDefinition(node) => Some(&node.src),
+            ASTNode::UsingForDirective(node) => Some(&node.src),
+            ASTNode::VariableDeclaration(node) => Some(&node.src),
+            ASTNode::VariableDeclarationStatement(node) => Some(&node.src),
+            ASTNode::WhileStatement(node) => Some(&node.src),
+        }
+    }
+}
