+++ src/detect/low/deprecated_oz_functions.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 
 use crate::{
-    context::loader::{ASTNode, ContextLoader},
+    context::{ast_node::ASTNode, loader::ContextLoader},
     detect::detector::{Detector, IssueSeverity},
 };
 use eyre::Result;
@@ -26,7 +26,8 @@ impl Detector for DeprecatedOZFunctionsDetector {
                     .absolute_path
                     .as_ref()
                     .map_or(false, |path| path.contains("openzeppelin"))
-            }) && identifier.name == "_setupRole" {
+            }) && identifier.name == "_setupRole"
+            {
                 self.found_deprecated_oz_functions
                     .push(Some(ASTNode::Identifier(identifier.clone())));
             }
@@ -43,7 +44,8 @@ impl Detector for DeprecatedOZFunctionsDetector {
                     .absolute_path
                     .as_ref()
                     .map_or(false, |path| path.contains("openzeppelin"))
-            }) && member_access.member_name == "safeApprove" {
+            }) && member_access.member_name == "safeApprove"
+            {
                 self.found_deprecated_oz_functions
                     .push(Some(ASTNode::MemberAccess(member_access.clone())));
             }
