+++ src/detect/medium/block_timestamp_deadline.rs
@@ -2,7 +2,7 @@ use std::error::Error;
 
 use crate::{
     ast::{Expression, FunctionCallKind},
-    context::loader::{ASTNode, ContextLoader},
+    context::{ast_node::ASTNode, loader::ContextLoader},
     detect::detector::{Detector, IssueSeverity},
 };
 use eyre::Result;
