+++ src/detect/nc/require_with_string.rs
@@ -2,7 +2,7 @@ use std::error::Error;
 
 use crate::{
     ast::Identifier,
-    context::loader::{ASTNode, ContextLoader},
+    context::{ast_node::ASTNode, loader::ContextLoader},
     detect::detector::{Detector, IssueSeverity},
 };
 use eyre::Result;
